[PW-8252] Code formatting
